+++ cmd/bdjuno/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/forbole/juno/v3/cmd"
 	initcmd "github.com/forbole/juno/v3/cmd/init"
@@ -16,8 +17,11 @@ import (
 	"github.com/forbole/bdjuno/v3/database"
 	"github.com/forbole/bdjuno/v3/modules"
 
-	osmosisapp "github.com/MonOsmosis/osmosis/v10/app"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	gammtypes "github.com/osmosis-labs/osmosis/v10/x/gamm/pool-models/balancer"
+
 	gaiaapp "github.com/cosmos/gaia/v7/app"
+	osmosisapp "github.com/osmosis-labs/osmosis/v10/app"
 )
 
 func main() {
@@ -66,6 +70,18 @@ func getBasicManagers() []module.BasicManager {
 // This should be edited by custom implementations if needed.
 func getAddressesParser() messages.MessageAddressesParser {
 	return messages.JoinMessageParsers(
+		OsmoMessageAddressesParser,
 		messages.CosmosMessageAddressesParser,
 	)
+}
+
+// OsmoMessageAddressesParser returns the list of all the accounts involved in the given
+// message if it's related to the x/gamm module
+func OsmoMessageAddressesParser(_ codec.Codec, cosmosMsg sdk.Msg) ([]string, error) {
+
+	if msg, ok := cosmosMsg.(*gammtypes.MsgCreateBalancerPool); ok {
+		return []string{msg.Sender, msg.FuturePoolGovernor}, nil
+	}
+
+	return nil, messages.MessageNotSupported(cosmosMsg)
 }