+++ modules/mint/source/remote/source.go
@@ -1,10 +1,10 @@
 package remote
 
 import (
-	minttypes "github.com/MonOsmosis/osmosis/v10/x/mint/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	mintsource "github.com/forbole/bdjuno/v3/modules/mint/source"
 	"github.com/forbole/juno/v3/node/remote"
+	minttypes "github.com/osmosis-labs/osmosis/v10/x/mint/types"
 )
 
 var (